+++ pkg/osquery/table/platform_tables_windows.go
@@ -4,9 +4,9 @@ package table
 
 import (
 	"github.com/kolide/launcher/pkg/osquery/tables/dataflattentable"
+	"github.com/kolide/launcher/pkg/osquery/tables/dsim_default_associations"
 	"github.com/kolide/launcher/pkg/osquery/tables/secedit"
 	"github.com/kolide/launcher/pkg/osquery/tables/wifi_networks"
-	"github.com/kolide/launcher/pkg/osquery/tables/windows_associations"
 	"github.com/kolide/launcher/pkg/osquery/tables/wmitable"
 
 	"github.com/go-kit/kit/log"
@@ -33,7 +33,7 @@ func platformTables(client *osquery.ExtensionManagerClient, logger log.Logger, c
 		ProgramIcons(),
 		secedit.TablePlugin(client, logger),
 		wifi_networks.TablePlugin(client, logger),
-		windows_associations.TablePlugin(client, logger),
+		dsim_default_associations.TablePlugin(client, logger),
 		wmitable.TablePlugin(client, logger),
 		dataflattentable.TablePluginExec(client, logger, "kolide_windows_updates",
 			dataflattentable.JsonType,
